+++ xdef/src/main/java/org/xdef/impl/parsers/XDParseGPS.java
@@ -5,6 +5,9 @@ import org.xdef.XDParserAbstract;
 import org.xdef.impl.code.DefGPSPosition;
 import org.xdef.msg.XDEF;
 import org.xdef.proc.XXNode;
+import org.xdef.sys.GPSPosition;
+import org.xdef.sys.SParser;
+import org.xdef.sys.SRuntimeException;
 
 /** Parse GPS value.
  * @author Vaclav Trojan
@@ -22,33 +25,59 @@ public class XDParseGPS extends XDParserAbstract {
 			if (p.isChar('(') && (p.isSignedFloat() || p.isSignedInteger())) {
 				double latitude =
 					Double.parseDouble(p.getParsedString().substring(1));
+				String name = null;
 				if (p.isChar(',')) {
 					int pos1 = p.getIndex();
 					if ((p.isSignedFloat() || p.isSignedInteger())) {
 						double longitude = Double.parseDouble(
 							p.getBufferPart(pos1, p.getIndex()));
+						double altitude = Double.MIN_VALUE;
 						if (p.isChar(',')) {
 							pos1 = p.getIndex();
 							if ((p.isSignedFloat() || p.isSignedInteger())){
-								double altitude = Double.parseDouble(
+								altitude = Double.parseDouble(
 									p.getBufferPart(pos1, p.getIndex()));
-								p.setParsedValue(new DefGPSPosition(
-									latitude, longitude, altitude));
+								if (p.isChar(',')) {
+									name = readGPSName(p);
+								}
+							} else {
+								name = readGPSName(p);
 							}
 						} else {
-							p.setParsedValue(
-								new DefGPSPosition(latitude, longitude));
+							if (p.isChar(',')) {
+								name = readGPSName(p);
+							}
 						}
 						if (p.isChar(')')) {
+							p.setParsedValue(new DefGPSPosition(new GPSPosition(
+								latitude, longitude, altitude, name)));
 							return;
 						}
 					}
 				}
 			}
-		} catch (Exception ex) {}
+		} catch (SRuntimeException ex) {
+			p.putReport(ex.getReport());
+		}
 		//Incorrect value of '&{0}'&{1}{: }
 		p.errorWithString(XDEF.XDEF809, parserName(),
 			p.getBufferPart(pos, p.getIndex()));
+	}
+
+	/** Read name of position. */
+	private String readGPSName(XDParseResult p) {
+		StringBuilder sb = new StringBuilder();
+		char ch;
+		p.isSpaces();
+		while ((ch = p.getCurrentChar()) != SParser.NOCHAR && ch != ')') {
+			sb.append(p.peekChar());
+		}
+		String result = sb.toString().trim();
+		if (result.isEmpty()) {
+			// Incorrect GPosition &{0}{: }
+			throw new SRuntimeException(XDEF.XDEF222, "name: " + result);
+		}
+		return result;
 	}
 
 	@Override